+++ _/app/index.js
@@ -1,0 +1,3 @@
+const span = document.createElement('span');
+span.innerText = 'bla';
+document.body.append(span);