+++ environment/app.webpack.config.js
@@ -1,0 +1,12 @@
+const path = require('path');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+module.exports = {
+    mode: 'development',
+    entry: './_/app/index.js',
+    output: {
+        path: path.resolve(__dirname, '../dist'),
+        filename: 'app.bundle.js',
+    },
+    plugins: [new HtmlWebpackPlugin()],
+};