+++ webpack.config.js
@@ -1,5 +1,7 @@
 const webworker = require('./environment/webworker.webpack.config.js');
+const app = require('./environment/app.webpack.config');
 
 module.exports = [
     webworker,
+    app,
 ];